Update pre-commit, new module snek5000.util.smake
