+++ .pre-commit-config.yaml
@@ -2,20 +2,20 @@
 # See https://pre-commit.com/hooks.html for more hooks
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v2.4.0
+  rev: v3.4.0
   hooks:
   - id: check-yaml
   - id: check-added-large-files
 - repo: https://github.com/psf/black
-  rev: stable
+  rev: 20.8b1
   hooks:
   - id: black
-- repo: https://github.com/timothycrosley/isort
-  rev: 4.3.21
+- repo: https://github.com/PyCQA/isort
+  rev: 5.8.0
   hooks:
   - id: isort
 - repo: https://gitlab.com/PyCQA/flake8
-  rev: 3.7.9
+  rev: 3.9.0
   hooks:
   - id: flake8
 - repo: https://github.com/snakemake/snakefmt
