+++ setup.cfg
@@ -64,7 +64,7 @@ dev =
     rope
     python-language-server
     fortran-language-server
-    pre-commit
+    pre-commit>=2.9.0
     twine
     readme-renderer[md]
